Move public job series functions into sarc.client: - load_job_series - compute_time_frames - compute_cost_and_waste
