+++ sarc/alerts/usage_alerts/cluster_scraping.py
@@ -5,8 +5,8 @@ from typing import List, Optional
 
 import pandas
 
+from sarc.client.series import compute_time_frames, load_job_series
 from sarc.config import MTL
-from sarc.jobs.series import compute_time_frames, load_job_series
 
 logger = logging.getLogger(__name__)
