+++ sarc/client/__init__.py
@@ -1,5 +1,6 @@
 from .gpumetrics import get_cluster_gpu_billings
 from .job import count_jobs, get_available_clusters, get_job, get_jobs
+from .series import compute_cost_and_waste, compute_time_frames, load_job_series
 from .users.api import get_user, get_users
 
 __all__ = [
@@ -10,4 +11,7 @@ __all__ = [
     "get_user",
     "get_users",
     "get_cluster_gpu_billings",
+    "load_job_series",
+    "compute_time_frames",
+    "compute_cost_and_waste",
 ]
